Add data property to any components that import data.js, and use that under the hood
