+++ src/components/index.js
@@ -5,3 +5,8 @@ export { default as Picker } from './picker'
 export { default as Preview } from './preview'
 export { default as Search } from './search'
 export { default as Skins } from './skins'
+export { default as NimbleCategory } from './nimble-category'
+export { default as NimbleEmoji } from './nimble-emoji'
+export { default as NimblePicker } from './nimble-picker'
+export { default as NimblePreview } from './nimble-preview'
+export { default as NimbleSearch } from './nimble-search'
