+++ src/utils/index.js
@@ -1,5 +1,4 @@
 import buildSearch from './build-search'
-import data from '../data'
 import stringFromCodePoint from '../polyfills/stringFromCodePoint'
 
 const _JSON = JSON
@@ -54,11 +53,7 @@ function sanitize(emoji) {
   }
 }
 
-function getSanitizedData() {
-  return sanitize(getData(...arguments))
-}
-
-function getData(emoji, skin, set) {
+function getData(emoji, skin, set, data) {
   var emojiData = {}
 
   if (typeof emoji == 'string') {
@@ -68,7 +63,7 @@ function getData(emoji, skin, set) {
       emoji = matches[1]
 
       if (matches[2]) {
-        skin = parseInt(matches[2])
+        skin = parseInt(matches[2], 10)
       }
     }
 
@@ -130,6 +125,10 @@ function getData(emoji, skin, set) {
   }
 
   return emojiData
+}
+
+function getSanitizedData() {
+  return sanitize(getData(...arguments))
 }
 
 function uniq(arr) {
